+++ src/Constraints/CreditCardValidator.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace SLLH\IsoCodesValidator\Constraints;
+
+use IsoCodes;
+use SLLH\IsoCodesValidator\AbstractIsoCodesConstraintValidator;
+use Symfony\Component\Validator\Constraint;
+
+/**
+ * Class CreditCardValidator.
+ *
+ * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ */
+class CreditCardValidator extends AbstractIsoCodesConstraintValidator
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function validate($value, Constraint $constraint)
+    {
+        parent::validate($value, $constraint);
+
+        if ($value && !IsoCodes\CreditCard::validate($value)) {
+            $this->createViolation($constraint->message);
+        }
+    }
+}