+++ src/IsoCodesConstraintValidator.php
@@ -5,9 +5,10 @@ namespace SLLH\IsoCodesValidator;
 trigger_error('SLLH\IsoCodesValidator\IsoCodesConstraintValidator is deprecated since version 1.1, to be removed in 2.0. Use SLLH\IsoCodesValidator\AbstractIsoCodesConstraintValidator instead.', E_USER_DEPRECATED);
 
 /**
- * Class IsoCodesConstraintValidator
+ * Class IsoCodesConstraintValidator.
  *
  * @author Sullivan Senechal <soullivaneuh@gmail.com>
+ *
  * @deprecated since version 1.1, to be removed in 2.0.
  *             Use {@link AbstractIsoCodesConstraintValidator} instead.
  */