+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * 1.1.0 (2015-??-??)
 
  * Deprecate `IsoCodesConstraintValidator` in favor of `AbstractIsoCodesConstraintValidator`.
+ * Add `CreditCard` constraint.
 
 * 1.0.3 (2015-06-02)